Module symbiosis::codegen

Structs

Doctype
A DOCTYPE token.

Line

Path

Struct

Structs

Text

Writer

Enums

Content
Types of text content.

Logic
Logic expressions.

Name

Scope
Representation of supported scope types.

StructName

Token

Type

Traits

Codegen
Code generators are used to generate template code in different languages and they have to implement this trait.

Implementations