Enum symbiosis::codegen::Scope

pub enum Scope {
    If(Logic),
    ForEach(Path, Tendril<UTF8, NonAtomic>, Option<Tendril<UTF8, NonAtomic>>),
}

Representation of supported scope types.

Variants

If

Everything within an if scope will be hidden if the provided logic expression is false.

ForEach

Repeat the content within the scope for each element in a collection. Args: collection, element, optional key