Trait symbiosis::codegen::Codegen

pub trait Codegen {
    fn build_template<W: Write>(&self, w: &mut W, name: &str, indent: u8, params: &HashMap<String, ContentType>, tokens: &[Token]) -> Result<(), Error>;

    fn build_module<W, F>(&self, w: &mut W, build_templates: F) -> Result<(), Error> where W: Write, F: FnOnce(&mut W, u8) -> Result<(), Error> { ... }
}

Code generators are used to generate template code in different languages and they have to implement this trait.

Required Methods

fn build_template<W: Write>(&self, w: &mut W, name: &str, indent: u8, params: &HashMap<String, ContentType>, tokens: &[Token]) -> Result<(), Error>

Generate code for a single template.

Provided Methods

fn build_module<W, F>(&self, w: &mut W, build_templates: F) -> Result<(), Error> where W: Write, F: FnOnce(&mut W, u8) -> Result<(), Error>

Generate code for a module or a similar collection containing multiple templates.